+++ test/device/index.js
@@ -275,6 +275,23 @@ describe('Device', function () {
         });
       });
 
+      describe('#macNormalized get', function () {
+        it('should return normalized mac from cached sysInfo', function () {
+          device.sysInfo.mac = 'My Test mac';
+          device.sysInfo.mic_mac = undefined;
+          device.sysInfo.ethernet_mac = undefined;
+          expect(device.macNormalized).to.eql('MYTESTMAC');
+          device.sysInfo.mac = undefined;
+          device.sysInfo.mic_mac = 'My Test mic_mac';
+          device.sysInfo.ethernet_mac = undefined;
+          expect(device.macNormalized).to.eql('MYTESTMICMAC');
+          device.sysInfo.mac = undefined;
+          device.sysInfo.mic_mac = undefined;
+          device.sysInfo.ethernet_mac = 'My Test ethernet_mac';
+          expect(device.macNormalized).to.eql('MYTESTETHERNETMAC');
+        });
+      });
+
       describe('#getSysInfo()', function () {
         it('should return info', function () {
           return expect(device.getSysInfo()).to.eventually.have.property('err_code', 0);